backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
transformers==4.35.2
torch==2.1.1
numpy<2.0.0
google-generativeai==0.3.1
groq==0.9.0
pydantic==2.5.0
python-multipart==0.0.6

